Update and rename README.md to Code
This is an innovative System for visually impaired people and acts as a voice assistant for them. This system is used to help the visually impaired to have access to the most important features of the phone enhancing the quality of the system making use of different custom layouts and using speech to text.